Fix linter issues and add config

+++ internal/raft/remote_test.go
@@ -59,9 +59,8 @@ func TestSetSnapshotAckWhenNotInSnapshotStateIsNotAllowed(t *testing.T) {
 }
 
 func TestRemoteString(t *testing.T) {
-	for _, tt := range []remoteStateType{remoteRetry,
-		remoteReplicate, remoteSnapshot} {
-		if tt.String() != remoteNames[uint64(tt)] {
+	for _, tt := range []remoteStateType{remoteRetry, remoteReplicate, remoteSnapshot} {
+		if tt.String() != remoteNames[tt] {
 			t.Errorf("unexpected string name")
 		}
 	}
